requirements.txt
pytest==8.3.3
selenium==4.25.0
allure-pytest==26.0.1
webdriver-manager==4.0.2
pyyaml-6.0.3
